Rename files and use bn over tokenAmount
