+++ src/components/uielements/coins/coinInputAdvanced/CoinInputAdvanced.style.ts
